fix: apply suggested changes
