+++ .github/workflows/deploy.yaml
@@ -83,7 +83,7 @@ jobs:
           fetch-depth: 0
 
       - name: Pre-generate sitemap lastmod manifest
-        run: python3 webapp/lastmod_manifest.py generate-lastmod
+        run: python3 scripts/generate-lastmod-manifest.py generate-lastmod
 
       - name: Use Node.js
         uses: actions/setup-node@v4
